Moved CooldownCreationListener to the "cooldown" package
