+++ src/main/java/dte/cooldownsystem/cooldown/Cooldown.java
@@ -18,7 +18,6 @@ import java.util.stream.Collectors;
 import org.apache.commons.lang.Validate;
 import org.bukkit.entity.Player;
 
-import dte.cooldownsystem.cooldown.listeners.CooldownCreationListener;
 import dte.cooldownsystem.cooldownfuture.CooldownFuture;
 
 /**
