+++ src/main/java/dte/cooldownsystem/cooldown/CooldownCreationListener.java
@@ -1,6 +1,4 @@
-package dte.cooldownsystem.cooldown.listeners;
-
-import dte.cooldownsystem.cooldown.Cooldown;
+package dte.cooldownsystem.cooldown;
 
 @FunctionalInterface
 public interface CooldownCreationListener
